web/FastPan/requirements.txt
# pywebview
aiosqlite
fastapi
SQLAlchemy
uvicorn
rich
pyyaml
httpx
regex
json5
objprint
nest_asyncio
python-multipart
aiofiles
pyjwt